% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd10.R
\name{get_icd10}
\alias{get_icd10}
\title{search strings for matches, returning data frame with associated ICD10 codes and PheCodes}
\usage{
get_icd10(cvec)
}
\arguments{
\item{cvec}{character()}
}
\value{
list named with elements of `cvec`; 
component i is a data.frame of matches of `cvec[i]`
in `phe.icd10$ICD10.String` using `grepl`
}
\description{
search strings for matches, returning data frame with associated ICD10 codes and PheCodes
}
\note{
uses data(phe.icd10) from phecopack
}
\examples{
nn = get_icd10(c("asthma", "malignant", "cardiac"))
sapply(nn, nrow)
lapply(nn, function(x) head(unique(x$ICD10.String)))
lapply(nn, function(x) head(unique(x$Phenotype)))
}
