% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phe.icd10}
\alias{phe.icd10}
\title{data.frame of ICD10, ICD10.String, PheCode, Phenotype, and Excl fields}
\format{
An object of class \code{data.frame} with 9505 rows and 6 columns.
}
\usage{
phe.icd10
}
\description{
data.frame of ICD10, ICD10.String, PheCode, Phenotype, and Excl fields
}
\note{
retrieved from \url{https://phewascatalog.org/phecodes_icd10} on May 6 2024, listed as v1.2 beta.  See \url{https://pubmed.ncbi.nlm.nih.gov/31553307/}
}
\keyword{datasets}
