% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecoapp.R
\name{phecoapp}
\alias{phecoapp}
\title{app for exploring mappings}
\usage{
phecoapp()
}
\description{
app for exploring mappings
}
\note{
Work in progress.  Intent was to allow a 
number of regexps to generate tables in tabs.
Currently just one table shown for one ICD10.String value.
}
